WIP: Add custom time stamp type that encodes like time.Time to JSON
This will enable prettier oasis-node control status CLI command output
for runtimes' latest time field.

+++ go/roothash/api/api.go
@@ -459,7 +459,7 @@ func SanityCheckBlocks(blocks map[common.Namespace]*block.Block) error {
 	for _, blk := range blocks {
 		hdr := blk.Header
 
-		if hdr.Timestamp > uint64(time.Now().Unix()+61*60) {
+		if uint64(hdr.Timestamp) > uint64(time.Now().Unix()+61*60) {
 			return fmt.Errorf("roothash: sanity check failed: block header timestamp is more than 1h1m in the future")
 		}
 	}
